requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-jose[cryptography]
bcrypt
transformers
torch
python-dotenv
twilio
openai
groq
better-profanity
